feat: integrate CIP labeling and harden operation result contracts
Add source-backed CIP graph/rule and label assignment implementations, typed e/z descriptors, detached atrop carrier queries, and registered public CIP operations. Move shared bond-direction stereo primitives into core and extend SMILES stereo/CX lowering and CX parser regression coverage.

Generate restricted pending-molecule result conversion through runtime validation and commit; strengthen capability privacy, transaction ownership, metadata, and failure-atomicity tests. Restore dimension-specific coordinates_2d/conformers_3d accessors and synchronize binding contracts and API documentation.

Enable full by default and wire all domain dependencies through capability features. Update crate documentation, Python num-bigint to 0.5.1, cc to 1.4.6 in the lockfile, and add SVG identity regressions requiring the kit.cosmol.org namespace.

Validation: parent strict cargo check passed; release strict runtime/default-feature/CIP API tests passed (87); SMILES CX lowering tests passed (11); formatting and staged whitespace checks passed. This is a progress snapshot, not full workspace acceptance: the unused smiles-to-stereo dependency awaits cleanup, and SVG identity tests still fail against installed Python 0.3.0. Exclude the three local architecture/plan/recovery Markdown files and ignored migration reports.

+++ crates/cosmolkit-cx/src/scan.rs
@@ -157,10 +157,16 @@ pub(crate) fn read_text_to(
                 ));
             }
             if next > entity_start + 2 {
+                // The pinned target has eight-bit signed `char`: RDKit first
+                // parses the full signed-int domain and its cast then retains
+                // the low byte (for example, 256 -> NUL and 321 -> `A`).
+                // `String` cannot hold RDKit's raw 0x80..=0xff bytes, so the
+                // detached text model lifts that byte reversibly to U+0080..=
+                // U+00FF instead of rejecting it or creating invalid UTF-8.
                 let value = text[entity_start + 2..next]
-                    .parse::<u8>()
+                    .parse::<i32>()
                     .map_err(|_| CxParseError::new(entity_start, "invalid CX character code"))?;
-                result.push(char::from(value));
+                result.push(char::from(value as u8));
             }
             *cursor = next + 1;
             segment_start = *cursor;
